Extract `EnsureFPointSliceCapacityIsPowerOfTwo` / `EnsureHPointSliceCapacityIsPowerOfTwo` to `types` package
