+++ pkg/streamingpromql/types/pool.go
@@ -3,8 +3,11 @@
 package types
 
 import (
+	"math/bits"
+
 	"github.com/prometheus/prometheus/promql"
 
+	"github.com/grafana/mimir/pkg/util/limiter"
 	"github.com/grafana/mimir/pkg/util/pool"
 )
 
@@ -26,4 +29,54 @@ func GetMatrix(size int) promql.Matrix {
 
 func PutMatrix(m promql.Matrix) {
 	matrixPool.Put(m)
+}
+
+// EnsureFPointSliceCapacityIsPowerOfTwo returns d if its capacity is already a power of two, or otherwise a new slice with the same elements and a
+// capacity that is a power of two.
+//
+// If a new slice is created, the memory consumption estimate is adjusted assuming the old slice is no longer used.
+//
+// This exists because many places in MQE assume that slices have come from our pools and always have a capacity that is a power of two.
+// For example, the ring buffer implementations rely on the fact that slices have a capacity that is a power of two.
+func EnsureFPointSliceCapacityIsPowerOfTwo(points []promql.FPoint, memoryConsumptionTracker *limiter.MemoryConsumptionTracker) ([]promql.FPoint, error) {
+	if pool.IsPowerOfTwo(cap(points)) {
+		return points, nil
+	}
+
+	nextPowerOfTwo := 1 << bits.Len(uint(cap(points)-1))
+	newSlice, err := FPointSlicePool.Get(nextPowerOfTwo, memoryConsumptionTracker)
+	if err != nil {
+		return nil, err
+	}
+
+	newSlice = newSlice[:len(points)]
+	copy(newSlice, points)
+
+	// Don't return the old slice to the pool, but update the memory consumption estimate.
+	// The pool won't use it because it's not a power of two, so there's no point in calling Put() on it.
+	memoryConsumptionTracker.DecreaseMemoryConsumption(uint64(cap(points))*FPointSize, limiter.FPointSlices)
+
+	return newSlice, nil
+}
+
+// EnsureHPointSliceCapacityIsPowerOfTwo is like EnsureFPointSliceCapacityIsPowerOfTwo, but for HPoint slices.
+func EnsureHPointSliceCapacityIsPowerOfTwo(points []promql.HPoint, memoryConsumptionTracker *limiter.MemoryConsumptionTracker) ([]promql.HPoint, error) {
+	if pool.IsPowerOfTwo(cap(points)) {
+		return points, nil
+	}
+
+	nextPowerOfTwo := 1 << bits.Len(uint(cap(points)-1))
+	newSlice, err := HPointSlicePool.Get(nextPowerOfTwo, memoryConsumptionTracker)
+	if err != nil {
+		return nil, err
+	}
+
+	newSlice = newSlice[:len(points)]
+	copy(newSlice, points)
+
+	// Don't return the old slice to the pool, but update the memory consumption estimate.
+	// The pool won't use it because it's not a power of two, so there's no point in calling Put() on it.
+	memoryConsumptionTracker.DecreaseMemoryConsumption(uint64(cap(points))*HPointSize, limiter.HPointSlices)
+
+	return newSlice, nil
 }
